Move security services into a new package
